refactor: props 를 컴포넌트 내부로 이동
